Removed rollover_worker
